Never use attendeesstring props to find attendees
The toattendeestring and ccattendeestring props contain the display
names of the attendees. Ambiguous display names (users having the same
display name) cannot be resolved to users.

It also doesn't seem necessary as the meeting request in a delegate's
inbox seems to be a full copy of the original meeting request
including recipients, and not a forwarded version of the meeting
request.

Still I admit that the meeting request code and functionality is pretty
hard to understand and I wouldn't be surprised if this commit does not
fix everything but instead breaks some cases.

Ref: KW-2798

related: https://jira.zarafa.com/browse/ZCP-7163 introduced the attendee
string functionality